requirements.txt
fastapi[all]
python-dotenv
uvicorn[standard]
pydantic
pydantic-settings
celery
flower
redis
gevent
requests
python-multipart
pymongo
python-jose[cryptography]
passlib[bcrypt]